DjangoCodeTest/requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
bootstrap4==0.1.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.1
defusedxml==0.7.1
Django==4.2.2
django-allauth==0.54.0
django-bootstrap4==23.1
django-forms-bootstrap==3.1.0
django-rest-swagger==2.2.0
djangorestframework==3.14.0
drf-yasg==1.21.7
drf-yasg2==1.19.4
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.3
oauthlib==3.2.2
openapi-codec==1.3.2
packaging==23.1
pycparser==2.21
PyJWT==2.7.0
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
ruamel.yaml==0.17.32
ruamel.yaml.clib==0.2.7
simplejson==3.19.1
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.3